+++ tests/unit/ModuleTest.php
@@ -9,8 +9,8 @@ namespace ymaker\email\templates\tests\unit;
 
 use Yii;
 use ymaker\email\templates\Module;
-use ymaker\email\templates\services\EmailTemplateService;
-use ymaker\email\templates\services\ServiceInterface;
+use ymaker\email\templates\repositories\EmailTemplatesRepository;
+use ymaker\email\templates\repositories\EmailTemplatesRepositoryInterface;
 use motion\i18n\LanguageProviderInterface;
 
 /**
@@ -24,7 +24,11 @@ class ModuleTest extends TestCase
     public function testInitService()
     {
         $module = new Module('test', null, ['languageProvider' => []]);
-        $this->assertInstanceOf(EmailTemplateService::class, Yii::$container->get(ServiceInterface::class));
+
+        $this->assertInstanceOf(
+            EmailTemplatesRepository::class,
+            Yii::$container->get(EmailTemplatesRepositoryInterface::class)
+        );
     }
 
     public function testInitLanguageProvider()
@@ -37,7 +41,7 @@ class ModuleTest extends TestCase
     {
         $module = new Module('test', null, ['languageProvider' => []]);
 
-        $this->assertTrue(Yii::$container->has(ServiceInterface::class));
+        $this->assertTrue(Yii::$container->has(EmailTemplatesRepositoryInterface::class));
         $this->assertTrue(Yii::$container->has(LanguageProviderInterface::class));
     }
 }